requirements.txt
genanki>=0.13
pandas>=2.0
gTTS>=2.4
Pillow>=10.0
pyyaml>=6.0
openai>=1.0.0
python-Levenshtein>=0.26.0
